Removed extra code for bomb spawning
Made the Bomb class extend from projectile controller. I wanted the bomb
to be tossed which is what the projectile controller does. Created its own
FixedUpdate that only moves it into position. It doesn't also handle
despawning like in projectile controller. Because of this, I could remove
the moving field from useAbility. This was only to get around the fact
that useAbility wanted a projectileController to work.