Started ability effects
Started implementing the different effects that abilities can have (stun,
push, pull, etc.).

Made push and pull effects. Sets a bool on the enemy to move towards a
specific position (the position they would be in after the push/pull).
This will probably break when hitting a wall, because it only stops when
the enemy gets to this new location. The enemy moves faster when being
pushed/pulled. Also is laggy when the enemy is in the middle of an attack
because it checks for the movement next frame AFTER the attack code. So if
the enemy is attacking, the push/pull code will get skipped until it isn't
attacking. Enemy logic needs to be redone.
Stun effect right now just stops the enemy from moving. Enemy shouldn't
even be able to be pushed while stunned. This needs more work, just did
this to test if getting stunned, waiting 2 sec, then getting unstunned
worked and it does.

Made a custom editor for the abilities. Abilities can now swap between the
different status effects (push, pull, stun, etc.). This also hides/shows
the appropriate variables (push distance when push is selected,
stunDuration when stun is selected).
Damage and cooldown are shown for all abilities regardless of additional
effects.
Only 1 additional effect per ability so far.
additional effects are only wiried up for aoeAbility at the moment.

AoEPullAbility class is now useless.

Added an incombat mode for the AI. Decisions are made faster while in
combat. This is to combat the lag when an ability hits an enemy and they
don't react. Currently, the only way for enemies to react is when the
player enters their aggro range.
The enemy still has no way of getting into combat yet (inCombat is always
false).