Damage over time
Implemented the damage over time effect.
Abilities can do damage over time. They can set the damage per tick, time
between ticks and total time.
Inspector works for damage over time.

Projectile abilities should also be able to use the status effects now.

AoE radius is now a variable in AoeDamageAbility and not a part of the
status effect (pushScale or pullScale).

Changed the inspector to be able to show all the other public variables.
Only the variables pertaining to the additional effects are hidden when
not needed. This works for classes that extend Ability that have their own
specific fields. No extra code needs to be written for them. Only extra
code needs to be written for the status effects.