Slows
Created a slow status effect. It slows the move speed of the enemies. Can
change the slow percent (0% to 100%) and the duration of the slow.
Can also have a decaying slow (recover to normal move speed over the
duration of the slow instead of a linear slow percent. But this doesn't
work yet.
First version only worked for one slow at a time. Fixed by doing the
following:
Created a slow struct that stores the percent and the end time of the
slow.
Created a list to store the slows on the enemy. Slows get added to the
list when they happen. The list gets sorted with the largest percent slows
at the start. The largest slow (first in the list) affects the enemy
movement. The slows are checked every frame to see if they've ended. If
they've ended, the next slow in the list is used (unless it has ended
too).

In the custom editor:
Used the original style for displaying the fields. The new style I was
using would only take effect once I clicked "Apply" in the inspector. The
old way takes effect when you press enter or leave the field focus.

When the enemy dies, the stuns and slows on it end.

renamed dotDamageinterval to dotTimeInterval
Removed some old comments. Wrote better comments on the attacking sequence.

Known bugs:
Multiple stuns don't work. Getting hit by a second stun during the first
will do nothing.
None of these status effects work on the player yet, only on enemies. All
the code will need to be moved from EnemyController to somewhere else.
Need to click "Apply" after changing decayingSlow in the inspector (not
that decaying slow does anything at the moment...)