Reworked Movement
Added a new class: MovementController. This is now the parent class of
anything that wants to move. Namely, the player and enemies.
Moved all movement code into this class instead of having it in
PlayerController and EnemyController.

Made this class so that the player could be affected by the same cc
effects that the enemy experiences.

Added another new class: MovementAbility. This is intended to be the class
to handle movement abilities like dashes. Does nothing at this time.

The player's dash is now broken. The functionality of it will be remade in
MovementController and MovementAbility so it can be used in multiple
places and so enemies can dash.
Clearing stuns and slows on death is broken currently. Stuns and roots are
probably not cleared on death.