Dash Ability
Made an ability that makes the entity dash a distance towards the mouse.
It just uses the displacement method that pushes and pulls use.

The movementAbility also has an aoeDamageAbility as a child so that it can
collide with things and do damage or cc things that it dashes through.

Added a variable to aoeDamageAbility that controls the explosion timing.
Used to be the explosion was enabled for 0.2 sec, now it can be whatever.
So the explosion survives the whole length of the dash.

Changed abilities called by the player to all use useAbility(spawnPoint,
aimPoint) even if they don't need those variables so that they can be
dragged and dropped without having to change the code.

Fixed a logic error when trying to stop the player from moving while
stunned or rooted, but still allowing them to move when being displaced.
Because the dash uses the displacement code of the pushes and pulls, this
means that the player can still dash while cced. Not sure if I want this
to happen or not. Porbably not.