Map Generation and Pathfinding
Added some procedural map generation from an online tutorial. Recreated A*
from wikipedia as pathfinding.

Astar works with the map generation. I checks the map for areas that are
walls and doesn't let the path go into walls.
The heuristic and cost between two points may not be working properly, but
they seem to work well enough to return a path.

The materials are for the floor of the map and the walls. There is support
for walls with textures in the meshGen, but I don't have a texture to put
in it.

Player prefab renamed some children objects.

Map Scene is a new scene for testing the map generation.
The map doesn't spawn any entities like players or enemies yet.

MovementController has been modified to try to allow path finding in the
AI movement.
Astar can return a path and move an object along the path, but that logic
hasn't been added into the enemyAI logic yet.