Pathfinding Optimization and Entity Spawning in Map
Optimized A* using a heap for the openSet instead of a list. This makes
finding the item with the lowest fScore O(1) instead of O(n) for searching
through a list.
Switched to a hashset for the closedSet. closedSet is only checked to see
if an item already exists in the set. hashSet makes this O(1) instead of
O(n) for searching a list.
Changed many helper methods to use AstarNode instead of Vector3 or Coord

To use the heap, I needed something to go in the heap. This was AstarNode.
It stores the grid position, the world position, fScore, and gScore of a
node on the graph.

The player and enemies can be spawned around the map.
The player is spawned at a specified spawnPoint. A room is carved out
around that point before the smoothing. This should ensure the room is
still there at the end of smooting and the process of filling in small
rooms (not very thouroughly tested yet).
Enemies are spawned at random points around the map that aren't wall and
aren't too near the player.

CameraController, ChildCollision, EnemyAI, EnemyController,
MovementController moved some assignment from start to update or other
methods
When spawning around the map, need to pass a reference to the map for
pathfinding. Also need to give them a reference to the player. Used to be
able to find the player in their start(), but the player is spawned in too
so they need to check for him again in update() til they find the player.

Gave enemy AI an anger counter and memory time. When they take damage,
they get angry. This keeps them in combat. After not taking damage for
memory time, they leave combat.
This helps the condition where the player could damage them from outside
their aggro range. Because the player was never close enough, they would
never chase.

Gave pathfinding to the enemyController. Now the enemy should path to the
player if the player is close enough, but they can't see him (through a
wall). Also paths when the enemy is in comabt from taking damage.
(currently always assumes the player did the damage regardless of source,
but at the moment, the player is the only possible source of damage)
Enemy will pathfind back to spawn if they leave combat away from the spawn
point.
Enemy will only pathfind if it can't see its current target. Once it can
see its target, it moves straight towards it. (Typing this out now, I
realize this may be a bad asssumption. The enemy may want to path around
some hazards like fire on the ground instead of moving straight to its
goal).

Added support for scaling the map (partially)
The map can scale in the x and z to be bigger. Still makes the same map,
it doesn't make a map with more points in it (to do that, change the width
in the mapGen script).
Returns the proper points for pathfinding and spawning purposes.

Known issues:
moving straight to target instead of pathing
Enemies may not always look where they're going (usually while
following a path)
Enemy assumes player did the damage and moves towards player's current
position, not towards position of damage)
Map scaling in y doesn't spawn entities at proper height
Enemy dying and respawning doesn't work 100%. Enemies can keep some
previous state. Ususally stays in comabt and immediately runs straight
towards the player when they respawn. Likely caused when the enemy dies in
the middle of a function call.