Melee Ability, New Player Class
Added the Fire Forged, a new player class.
Has a left click, right click, and space ability

Created a melee ability class. It moves a collider from one point to
another. Thus is could be used for abilities other than melee abilities.
The melee enemy now has a similar attack to the melee ability. It doesn't
currently use the ability system, but its attack is pretty much just the
melee ability copy/pasted.

Updated the scale for the game.
Players and enemies are at 1 scale instead of 2.
Camera is moved slightly.
Abilities are slightly smaller.

Abilities now tell the caster when they finish. This means 2 abilities
can't be cast at the same time. One goes first, then you have to wait
until it finishes before another ability can be cast.
The player can't rotate their character while an ability is being cast.
To do: a similar solution for slowing the player down while casting
abilities, having more timings to the abilities to separate the ability
into cast times and windup and follow through times where applicable.

Melee ability and enemy's attack uses an IEnumerator for the attack. This
allows the ability to happen over time without having to change code in
the Update loop. This is an alternative to calling a chain of Invokes.
Abilities can be separated into cast times, windups, follow throughs, etc.

Created a method in MovementController for looking at your target.
Currently, it doesn't let an entity look while it's stunned. It could have
further logic in the future.

PlayerController:
Started to remove the shield bash code from the playerController class.
Going forward, the shield bash would be made into an ability and slotted
in. The PlayerController class should be agnostic of the specific
abilities.
On a related note, ArcherPlayerController is probably a redundant class
that should get deleted in lieu of using the base PlayerController class
with abilities making up the differences in classes.
Abilities check to make sure another ability isn't currently in progress
before being cast.
The StartAbility() method could possibly be done better. It seems like it
could cause problems if more keybinds get added. It also may not work if
the game later on supports action bars where you can click on abilities to
use them.
FixedUpdate has some logic to have the character look at the mouse. The
base MovementController class has its own LookAt() method that this class
maybe should be using instead.
StartAbility() has some room for error if new keybinds are added. It would
need to be adapted and could lead to bugs if not done correctly.

Ability:
Made an event that the player can subscribe to for when an ability
finishes. This event is called in each ability child class in a location
that may be appropriate.

Fixed a bug where the player could push against a wall and then have some
residual forces that push the character around.
The fix is to give the player's rigidbody a drag of 10. 10 was chosen
because it seemed to make the force dissapate quickly enough that it isn't
noticable.
Fixed a bug in the damage over time component of the health. It would
always do 1 tick too many in the DoT. Now it checks if the DoT has ended
before doing damage.
The DoT code only allows for 1 DoT at a time with new DoTs completely
removing the old DoT. This will need to be updated sometime.