Better SpawnAbility
Increased fidelity on SpawnAbility.
Now can fire multiple projectiles with one ability.
Can have a cast time before the projectile is fired
Caster can be slowed(or completely stopped) while the ability is being used
Subsequent projectiles can have individual timings
p1 fires, wait 0.3 seconds, p2 fires
multiple projectiles can be fired at the same time
Player can adjust aim between projectiles
Can lock aim direction for all projectiles (all projectiles will fire in
the same direction relative to the player's position even if the player
moves)

To do:
Give different projectiles different damage (all do the same damage
currently)
Allow variations in start and end point. Currently, all projectiles start
at the player and go to the aimPoint. Might be nice to have functionality
for a spread shot - projectiles fan out from the player.

Ability class
Has some new Actions (aimPermission and slowMovement)
slowMovement allows each ability to say when it would like the caster to
slow down (usually to cast an ability)
The caster will subscribe to these Actions and then handle it
aimPermission currently does nothing. Its intention was to stop the player
from rotating in the middle of abilities.

MovementController
Added a selfSlowPercent to slow the caster.
This is ignored while being displaced.
This is different from the slows that abilities can cause as it has no set
duration.
It is also multiplied by any active slows from other sources.
getAimpoint() was added so that an ability can assume that anything that
can move can also aim. This is what the player and enemies have in common.
Because of this, the ability doesn't need to account for differences in
players and enemies; the player and enemy classes worry about that.

PlayerController
SpawnAbility needs to know where the player is aiming when the cast time
finishes, not at the spart of the ability cast.
Melee ability locks the player's rotation while using an abilty so the
player can't turn to extend the hitbox of the melee ability.
This needed to be removed so SpawnAbility could work.
AMEND: Moved the if(!usingAbility) so it still calculates aimPoint (what
spawnAbility needs), but doesn't rotate the player (what meleeAbility
needs).
This solution isn't perfect (the player can aim between projectiles in a
multi-projectile ability, but the character won't rotate so it will look
weird).
The solution may be to utilize the aimPermission action that currently
does nothing.
Added the Action code for slowMovement to StartAbility where it already
initialized the other action, abilityOver