Removed Shield, Added Arrows
Removed shield code from the playercontroller.
If the player gets a shield later, it will have to be added in a way that
works with other abilities or expand the player to be able to handle
multiple weapons (sword and shield, dual swords, etc.)

EnemyController
Enemy would check to see if it could see the target (the player). When the
player fired an arrow at the enemy, it would no longer be able to see the
player (the arrow blocked its view).
Changed the canSeeTarget raycast to only cast on the target's layer and
the obstacle layer (walls).

ProjectileController
Added a flag to let the arrow destroy itself when it gets to its target
position. This is so the arrow doesn't travel to where you aimed only to
stop in mid-air and wait until it timed out. It now destroys itself
shortly after getting there (0.5 seconds).
Reaching the end turns off the collider. Switched from turning off the
sphereCollider to turning off a collider so the collider can be any shape
(arrow is a capsule).

Added an arrow prefab. Looks bad, but better than spheres.